Bugfix/YM-7561 fe android send money screen exc

+++ sendmoney/src/main/java/co/yap/sendmoney/fundtransfer/viewmodels/CashTransferViewModel.kt
@@ -165,7 +165,7 @@ class CashTransferViewModel(application: Application) :
                 transactionRepository.cashPayoutTransferRequest(
                     SendMoneyTransferRequest(
                         beneficiaryId = beneficiaryId,
-                        amount = state.amount.toDouble(),
+                        amount = state.amount,
                         currency = "AED",
                         purposeCode = "8",
                         remarks = state.noteValue
